Rename: AuthController로 파일 통합
